tp53_rag/requirements.txt
# ============================================================
# TP53 RAG Platform - Requirements
# Gemma 4 Hackathon | Health & Sciences Track
# ============================================================

# ── Structure Visualisation Agent (Agent #7) ──────────────
transformers>=4.40.0
torch==2.6.0
accelerate==0.30.0
fair-esm==2.0.0
umap-learn==0.5.6
py3Dmol==2.1.0
flask==3.0.3
flask-cors==4.0.1
plotly==5.22.0
ipywidgets==8.1.3

# Core RAG & Vector DB
chromadb==0.5.0
langchain==0.2.0
langchain-community==0.2.0
langchain-chroma==0.1.2
langchain-ollama==0.1.3

# Document Processing
pypdf==4.2.0
unstructured>=0.16.19
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2

# Embeddings (local, no API key needed)
sentence-transformers==3.0.1

# Bioinformatics
biopython==1.83

# Data handling
pandas==2.2.2
numpy>=2.0.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
rich==13.7.1
tqdm==4.66.4
loguru==0.7.2

# Testing
pytest==8.2.2
pytest-cov==5.0.0

# API serving (for n8n integration later)
fastapi==0.111.0
uvicorn==0.30.1
